voice/requirements.txt
sounddevice>=0.5
numpy>=2.2
pyyaml>=6.0
onnxruntime>=1.16
keyboard>=0.13
